Factor out FixedResultsSearchCommander
